+++ exploration-module/src/main/java/io/edpn/backend/exploration/adapter/persistence/SystemRepository.java
@@ -4,10 +4,10 @@ package io.edpn.backend.exploration.adapter.persistence;
 import io.edpn.backend.exploration.adapter.persistence.entity.MybatisSystemEntity;
 import io.edpn.backend.exploration.application.domain.System;
 import io.edpn.backend.exploration.application.dto.mapper.SystemEntityMapper;
-import io.edpn.backend.exploration.application.port.outgoing.CreateSystemPort;
-import io.edpn.backend.exploration.application.port.outgoing.LoadSystemPort;
-import io.edpn.backend.exploration.application.port.outgoing.LoadSystemsByNameContainingPort;
-import io.edpn.backend.exploration.application.port.outgoing.SaveSystemPort;
+import io.edpn.backend.exploration.application.port.outgoing.system.CreateSystemPort;
+import io.edpn.backend.exploration.application.port.outgoing.system.LoadSystemPort;
+import io.edpn.backend.exploration.application.port.outgoing.system.LoadSystemsByNameContainingPort;
+import io.edpn.backend.exploration.application.port.outgoing.system.SaveSystemPort;
 import io.edpn.backend.util.IdGenerator;
 import io.edpn.backend.util.exception.DatabaseEntityNotFoundException;
 import lombok.RequiredArgsConstructor;
