+++ exploration-module/src/main/java/io/edpn/backend/exploration/application/port/outgoing/system/LoadSystemsByNameContainingPort.java
@@ -1,4 +1,4 @@
-package io.edpn.backend.exploration.application.port.outgoing;
+package io.edpn.backend.exploration.application.port.outgoing.system;
 
 import io.edpn.backend.exploration.application.domain.System;
