+++ exploration-module/src/main/java/io/edpn/backend/exploration/application/port/outgoing/systemcoordinaterequest/DeleteSystemCoordinateRequestPort.java
@@ -1,4 +1,4 @@
-package io.edpn.backend.exploration.application.port.outgoing;
+package io.edpn.backend.exploration.application.port.outgoing.systemcoordinaterequest;
 
 import io.edpn.backend.util.Module;
