+++ exploration-module/src/main/java/io/edpn/backend/exploration/application/port/outgoing/topic/CreateTopicPort.java
@@ -1,4 +1,4 @@
-package io.edpn.backend.exploration.application.port.outgoing;
+package io.edpn.backend.exploration.application.port.outgoing.topic;
 
 import java.util.concurrent.CompletableFuture;
