+++ exploration-module/src/test/java/io/edpn/backend/exploration/application/service/SystemControllerServiceTest.java
@@ -5,7 +5,7 @@ import io.edpn.backend.exploration.application.domain.exception.ValidationExcept
 import io.edpn.backend.exploration.application.dto.SystemDto;
 import io.edpn.backend.exploration.application.dto.mapper.SystemDtoMapper;
 import io.edpn.backend.exploration.application.port.incomming.FindSystemsByNameContainingUseCase;
-import io.edpn.backend.exploration.application.port.outgoing.LoadSystemsByNameContainingPort;
+import io.edpn.backend.exploration.application.port.outgoing.system.LoadSystemsByNameContainingPort;
 import io.edpn.backend.exploration.application.validation.LoadByNameContainingValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
